Implement new NativeCryptoFacade methods on Android
- `hmacSha256()`
- `verifyHmacSha256()`
